feat(mcp): module
