+++ auto-configurations/spring-ai-alibaba-autoconfigure-mcp-recovery-client/src/main/resources/META-INF/spring-configuration-metadata.json
@@ -2,8 +2,8 @@
   "groups": [
     {
       "name": "spring.ai.alibaba.mcp.recovery",
-      "type": "com.alibaba.cloud.ai.autoconfigure.mcp.client.config.McpRecoveryAutoProperties",
-      "sourceType": "com.alibaba.cloud.ai.autoconfigure.mcp.client.config.McpRecoveryAutoProperties",
+      "type": "com.alibaba.cloud.ai.mcp.client.config.McpRecoveryProperties",
+      "sourceType": "com.alibaba.cloud.ai.mcp.client.config.McpRecoveryProperties",
       "description": "Configuration properties for MCP recovery client."
     }
   ],
@@ -12,21 +12,21 @@
       "name": "spring.ai.alibaba.mcp.recovery.enabled",
       "type": "java.lang.Boolean",
       "description": "Whether to enable MCP Recovery Client",
-      "sourceType": "com.alibaba.cloud.ai.autoconfigure.mcp.client.config.McpRecoveryAutoProperties",
+      "sourceType": "com.alibaba.cloud.ai.mcp.client.config.McpRecoveryProperties",
       "defaultValue": false
     },
     {
       "name": "spring.ai.alibaba.mcp.recovery.delay",
       "type": "java.time.Duration",
       "description": "Delay between MCP recovery attempts.",
-      "sourceType": "com.alibaba.cloud.ai.autoconfigure.mcp.client.config.McpRecoveryAutoProperties",
+      "sourceType": "com.alibaba.cloud.ai.mcp.client.config.McpRecoveryProperties",
       "defaultValue": "5s"
     },
     {
       "name": "spring.ai.alibaba.mcp.recovery.stop",
       "type": "java.time.Duration",
       "description": "Duration after which MCP recovery attempts will stop.",
-      "sourceType": "com.alibaba.cloud.ai.autoconfigure.mcp.client.config.McpRecoveryAutoProperties",
+      "sourceType": "com.alibaba.cloud.ai.mcp.client.config.McpRecoveryProperties",
       "defaultValue": "10s"
     }
   ],
