+++ auto-configurations/spring-ai-alibaba-autoconfigure-mcp-recovery-client/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -13,5 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-com.alibaba.cloud.ai.autoconfigure.mcp.client.config.McpRecoveryAutoConfiguration
+com.alibaba.cloud.ai.autoconfigure.mcp.client.McpRecoveryAutoConfiguration
