+++ spring-ai-alibaba-mcp/pom.xml
@@ -19,6 +19,7 @@
         <module>spring-ai-alibaba-mcp-nacos</module>
         <module>spring-ai-alibaba-mcp-nacos-common</module>
         <module>spring-ai-alibaba-mcp-gateway</module>
+        <module>spring-ai-alibaba-mcp-recovery</module>
     </modules>
     <scm>
         <connection>git://github.com/alibaba/spring-ai-alibaba.git</connection>
