+++ spring-ai-alibaba-mcp/spring-ai-alibaba-mcp-recovery/pom.xml
@@ -8,10 +8,11 @@
         <relativePath>../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>spring-ai-alibaba-autoconfigure-mcp-recovery-client</artifactId>
+    <artifactId>spring-ai-alibaba-mcp-recovery</artifactId>
 
-    <name>Spring AI Alibaba Mcp Recovery Client Autoconfiguration</name>
-    <description>Spring AI Alibaba Mcp Recovery Client Autoconfiguration</description>
+    <version>${revision}</version>
+    <name>Spring AI Alibaba MCP Recovery Client</name>
+    <description>MCP Recovery Client for Spring AI Alibaba</description>
     <url>https://github.com/alibaba/spring-ai-alibaba</url>
     <scm>
         <connection>git://github.com/alibaba/spring-ai-alibaba.git</connection>
@@ -19,17 +20,19 @@
         <url>https://github.com/alibaba/spring-ai-alibaba</url>
     </scm>
 
+    <properties></properties>
+
     <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
+        </dependency>
+
         <dependency>
             <groupId>org.springframework.ai</groupId>
             <artifactId>spring-ai-starter-mcp-client-webflux</artifactId>
         </dependency>
 
-        <dependency>
-            <groupId>com.alibaba.cloud.ai</groupId>
-            <artifactId>spring-ai-alibaba-mcp-recovery</artifactId>
-            <version>${revision}</version>
-        </dependency>
     </dependencies>
 
 </project>
