+++ spring-ai-alibaba-mcp/spring-ai-alibaba-mcp-recovery/src/main/java/com/alibaba/cloud/ai/mcp/client/component/McpReconnectTask.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.cloud.ai.autoconfigure.mcp.client.component;
+package com.alibaba.cloud.ai.mcp.client.component;
 
 import java.util.concurrent.Delayed;
 import java.util.concurrent.TimeUnit;
